# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venom/ontology/search:
    get:
      operationId: searchVenomOntology
      tags: [venom]
      summary: Search the signed-in user's knowledge ontology across projects
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: org
          in: query
          required: false
          description: >-
            Company id. When present, searches that company's shared Brain
            instead of the personal ontology; the caller must be a member.
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Concepts matching the query, best matches first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomOntologySearchResponse"
        "400":
          description: Invalid search query
        "401":
          description: Authentication required
  /venom/ontology/concepts/{conceptId}:
    get:
      operationId: getVenomOntologyConcept
      tags: [venom]
      summary: Get one ontology concept with its neighbors and evidence
      parameters:
        - name: conceptId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
        - name: org
          in: query
          required: false
          description: >-
            Company id. When present, resolves the concept inside that
            company's shared Brain; the caller must be a member.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      responses:
        "200":
          description: The concept, its linked neighbors, and its evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomOntologyConceptDetail"
        "401":
          description: Authentication required
        "404":
          description: Concept not found
  /venom/conversations/{conversationId}:
    get:
      operationId: getVenomConversation
      tags: [venom]
      summary: Read one conversation from the synced workspace snapshot
      description: >-
        Serves a cited conversation for read-only viewing on a device that
        does not hold it locally. Brain evidence on a server-side concept can
        reference conversations that only exist in the cloud snapshot; this
        lookup completes that trail of proof. Owner-scoped - it only ever
        reads the signed-in user's own snapshot.
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        "200":
          description: The conversation as last synced, labeled with its project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomRemoteConversation"
        "401":
          description: Authentication required
        "404":
          description: No synced conversation with this id
...
components:
...
  schemas:
...
    VenomResponseMode:
      type: string
      enum: [talk, verify, debate]
      description: >-
        How Venom answers — a single assistant (talk), background multi-voice
        verification that converges on one conclusion (verify), or a visible
        multi-voice debate in the thread (debate).
...
    VenomVoiceModelPick:
      type: object
      additionalProperties: false
      description: >-
        Assigns one deliberation voice to a specific managed model. Opposing
        voices must run on different LLM providers; the server rejects picks
        that would make a model argue itself.
      properties:
        voiceId:
          type: string
          enum: [direct, skeptic, evidence]
        modelId:
          $ref: "#/components/schemas/VenomModelId"
      required: [voiceId, modelId]
    VenomModelId:
      type: string
      enum: [venom-gpt, venom-claude, venom-gemini, venom-grok]
    VenomChatFileKind:
      type: string
      enum: [upload, generated]
      description: upload = host gave Venom a file; generated = Venom authored it.
...
    VenomMessageAttachment:
      type: object
      additionalProperties: false
      description: >-
        Compact attachment stamp carried on a chat message. The id resolves to
        a VenomChatFile; display metadata is denormalized so history renders
        without a lookup. Attachment stamps ride the synced workspace blob, so
        they must stay tiny.
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 160
        contentType:
          type: string
          minLength: 1
          maxLength: 100
        size:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 10485760
        kind:
          $ref: "#/components/schemas/VenomChatFileKind"
        thumbnail:
          type: string
          minLength: 1
          maxLength: 24000
          pattern: "^data:image/(jpeg|png|webp);base64,[A-Za-z0-9+/=]+$"
          description: >-
            Tiny inline preview (JPEG/PNG/WEBP data URL) for image
            attachments, generated client-side at attach time. Bounded hard
            because stamps ride the synced workspace blob; full image bytes
            never sync.
      required: [id, name, contentType, size, kind]
    VenomVoiceId:
      type: string
      enum: [direct, skeptic, evidence]
...
    VenomOntologySearchResult:
      type: object
      additionalProperties: false
      required:
        [
          id,
          projectId,
          label,
          category,
          summary,
          strength,
          mentionCount,
          lastUpdatedAt,
          evidenceCount,
        ]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 120
        projectId:
          type: string
          maxLength: 120
          nullable: true
        label:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          minLength: 1
          maxLength: 100
        summary:
          type: string
          maxLength: 2000
        strength:
          type: number
          minimum: 0
          maximum: 1
        mentionCount:
          type: number
          minimum: 0
        lastUpdatedAt:
          type: number
          minimum: 0
        evidenceCount:
          type: number
          minimum: 0
    VenomOntologySearchResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/VenomOntologySearchResult"
    VenomOntologyConceptDetail:
      type: object
      additionalProperties: false
      required: [concept, neighbors, people]
      properties:
        concept:
          $ref: "#/components/schemas/VenomKnowledgeCluster"
        neighbors:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/VenomOntologySearchResult"
        people:
          description: >-
            Identity labels for every person attributed to this concept's
            evidence. Evidence captured before attribution is reported under
            the ontology owner's identity, so every evidence entry's
            capturedByUserId resolves against this list.
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/VenomOntologyPerson"
    VenomOntologyPerson:
      type: object
      additionalProperties: false
      required: [userId, displayName]
      properties:
        userId:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          description: >-
            Best display label for the person (their name, else their
            email). Null when the identity record has neither.
          type: [string, "null"]
          maxLength: 320
...
    VenomDeliberationTake:
      type: object
      additionalProperties: false
      properties:
        voiceId:
          $ref: "#/components/schemas/VenomVoiceId"
        name:
          type: string
          minLength: 1
          maxLength: 80
        modelId:
          $ref: "#/components/schemas/VenomModelId"
        modelName:
          type: string
          minLength: 1
          maxLength: 80
        content:
          type: string
          maxLength: 8000
        status:
          type: string
          enum: [ok, failed]
      required: [voiceId, name, content, status]
    VenomMessageDeliberation:
      type: object
      additionalProperties: false
      properties:
        voices:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/VenomDeliberationTake"
        disagreements:
          type: array
          maxItems: 8
          items:
            type: string
            maxLength: 500
      required: [voices, disagreements]
    VenomMessage:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 120
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          maxLength: 50000
        createdAt:
          type: number
          multipleOf: 1
          minimum: 0
        status:
          type: string
          enum: [sending, sent, error]
        modelId:
          $ref: "#/components/schemas/VenomModelId"
        modelName:
          type: string
          minLength: 1
          maxLength: 80
        deliberation:
          $ref: "#/components/schemas/VenomMessageDeliberation"
        speakerId:
          type: string
          minLength: 1
          maxLength: 64
          description: >-
            Debate voice identity behind this assistant turn. Absent for
            ordinary replies, so older clients render debate turns as plain
            assistant messages.
        speakerName:
          type: string
          minLength: 1
          maxLength: 80
          description: Display name of the debate voice behind this assistant turn.
        attachments:
          type: array
          maxItems: 5
          description: >-
            Files carried by this message — uploads on user turns, generated
            files on assistant turns. Optional, so older clients and existing
            history remain valid.
          items:
            $ref: "#/components/schemas/VenomMessageAttachment"
      required: [id, role, content, createdAt, status]
    VenomConversation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 120
        title:
          type: string
          minLength: 1
          maxLength: 200
        projectId:
          type: [string, "null"]
          maxLength: 120
        updatedAt:
          type: number
          multipleOf: 1
          minimum: 0
        messages:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/VenomMessage"
        responseMode:
          $ref: "#/components/schemas/VenomResponseMode"
        blend:
          $ref: "#/components/schemas/VenomConversationBlend"
        modeUpdatedAt:
          type: number
          multipleOf: 1
          minimum: 0
          description: >-
            When the response mode, blend, or voice picks last changed on any
            device; the newer block wins in cross-device merges.
        voiceModels:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/VenomVoiceModelPick"
          description: >-
            Per-conversation verify voice assignments. Part of the response-mode
            preference block: it merges on modeUpdatedAt with mode and blend so
            a pick made on one device survives another device's snapshot merge.
      required: [id, title, projectId, updatedAt, messages]
    VenomConversationBlend:
      type: object
      additionalProperties: false
      description: >-
        The blend pad position for this conversation, stored as the three
        corner identities and their normalized weights.
      properties:
        corners:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: string
            minLength: 1
            maxLength: 64
        weights:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: number
            minimum: 0
            maximum: 1
      required: [corners, weights]
    VenomRemoteConversation:
      type: object
      additionalProperties: false
      description: >-
        A conversation served from the synced workspace snapshot for
        read-only viewing. projectName is resolved against the snapshot's own
        project list because the requesting device may not hold the project
        either. Heavy per-message internals (deliberation transcripts) are
        omitted; the payload is the readable exchange.
      properties:
        conversation:
          $ref: "#/components/schemas/VenomConversation"
        projectName:
          type: [string, "null"]
          maxLength: 200
      required: [conversation, projectName]
    VenomKnowledgeSource:
      type: object
      additionalProperties: false
      properties:
        conversationId:
          type: string
          minLength: 1
          maxLength: 120
        projectId:
          type: [string, "null"]
          maxLength: 120
        conversationTitle:
          type: string
          minLength: 1
          maxLength: 200
        messageIds:
          type: array
          maxItems: 12
          items:
            type: string
            minLength: 1
            maxLength: 120
        excerpt:
          type: string
          maxLength: 2000
        updatedAt:
          type: number
          multipleOf: 1
          minimum: 0
        capturedByUserId:
          description: >-
            Clerk user id of the account whose chat produced this evidence.
            Null for knowledge captured before attribution existed; renderers
            then attribute it to the ontology owner. The server discards
            stamps that name anyone other than the owner when absorbing
            client snapshots, so the value is trustworthy.
          type: [string, "null"]
          maxLength: 120
        capturedAt:
          description: >-
            When the capture that produced this evidence was filed, in ms
            since epoch. Null for pre-attribution evidence; renderers then
            fall back to updatedAt.
          type: [number, "null"]
          multipleOf: 1
          minimum: 0
        sensitive:
          description: >-
            Sensitivity lock on this evidence entry. Server-managed on
            workspace-tier knowledge: set only through the sensitivity
            endpoints, never through client snapshots. Locked evidence stays
            visible to members but is withheld from exports when the
            workspace's export policy forbids sensitive content leaving.
          type: boolean
      required:
        [conversationId, projectId, conversationTitle, messageIds, excerpt, updatedAt]
    VenomKnowledgeCluster:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 120
        projectId:
          type: [string, "null"]
          maxLength: 120
        label:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          minLength: 1
          maxLength: 100
        strength:
          type: number
          minimum: 0
          maximum: 1
        x:
          type: number
        y:
          type: number
        links:
          type: array
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 120
        description:
          type: string
          maxLength: 2000
        summary:
          type: string
          maxLength: 2000
        mentionCount:
          type: number
          multipleOf: 1
          minimum: 0
        lastUpdatedAt:
          type: number
          multipleOf: 1
          minimum: 0
        sources:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/VenomKnowledgeSource"
        sensitive:
          description: >-
            Sensitivity lock on the whole cluster. Server-managed on
            workspace-tier knowledge: set only through the sensitivity
            endpoints, never through client snapshots. Locked clusters stay
            visible to members but are withheld from exports when the
            workspace's export policy forbids sensitive content leaving.
          type: boolean
        adminOnly:
          description: >-
            Admin-only restriction on the whole cluster. Set only by
            workspace admins through the restriction endpoint. Restricted
            clusters are filtered out of member reads, member chat context,
            member citations, and member exports server-side, so a non-admin
            response never carries a restricted cluster at all — when this
            flag is true, the reader is an admin.
          type: boolean
        unsorted:
          description: >-
            Author-private Unsorted holding state, used when scope
            classification was not confident enough to file the cluster.
            Lives only on personal-tier records: it syncs across the
            author's own devices but never appears in workspace reads or
            workspace exports. Cleared when the concept is clarified —
            by a confident re-extraction, an automatic re-file, or the
            author sorting it by hand.
          type: boolean
      required:
        [id, projectId, label, category, strength, x, y, links, summary, mentionCount, lastUpdatedAt, sources]
...